scaladsl for CassandraSession, #1
* structure build for persistence-cassandra-scaladsl
* moved CassandraSessionImpl to akka-persistence-cassandra,
  since we already had a similar class there and some users
  of akka-persistence-cassandra have requested it
* akka-persistence-cassandra provides both scaladsl and javadsl
  and in lagom we can have thin api wrappers (delegating)
