+++ docs/manual/guide/cluster/PersistentEntityCassandra.md
@@ -22,4 +22,4 @@ In sbt:
 
 ## Configuration
 
-Lagom's Cassandra support is provided by the [`akka-persistence-cassandra`](https://github.com/akka/akka-persistence-cassandra) plugin. A full configuration reference can be in the plugins [`reference.conf`](https://github.com/akka/akka-persistence-cassandra/blob/v0.17/src/main/resources/reference.conf).
+Lagom's Cassandra support is provided by the [`akka-persistence-cassandra`](https://github.com/akka/akka-persistence-cassandra) plugin. A full configuration reference can be in the plugins [`reference.conf`](https://github.com/akka/akka-persistence-cassandra/blob/v0.19/src/main/resources/reference.conf).
