+++ docs/manual/guide/cluster/PubSub.md
@@ -72,4 +72,4 @@ The published messages must be serializable since they will be sent across the n
 
 ## Underlying Implementation
 
-It is implemented with [Akka Distributed Publish Subscribe](http://doc.akka.io/docs/akka/2.4.4/java/distributed-pub-sub.html).
+It is implemented with [Akka Distributed Publish Subscribe](http://doc.akka.io/docs/akka/2.4/java/distributed-pub-sub.html).
