+++ docs/manual/guide/cluster/ReadSideCassandra.md
@@ -96,5 +96,5 @@ Once you have finished registering all your event handlers, you can invoke the `
 
 The `CassandraSession` is using the [Datastax Java Driver for Apache Cassandra](https://github.com/datastax/java-driver).
 
-Each `ReadSideProcessor` instance is executed by an [Actor](http://doc.akka.io/docs/akka/2.4.4/java/untyped-actors.html) that is managed by [Akka Cluster Sharding](http://doc.akka.io/docs/akka/2.4.4/java/cluster-sharding.html). The processor consumes a stream of persistent events delivered by the `eventsByTag` [Persistence Query](http://doc.akka.io/docs/akka/2.4.4/java/persistence-query.html) implemented by [akka-persistence-cassandra](https://github.com/akka/akka-persistence-cassandra). The tag corresponds to the `tag` defined by the `AggregateEventTag`.
+Each `ReadSideProcessor` instance is executed by an [Actor](http://doc.akka.io/docs/akka/2.4/java/untyped-actors.html) that is managed by [Akka Cluster Sharding](http://doc.akka.io/docs/akka/2.4/java/cluster-sharding.html). The processor consumes a stream of persistent events delivered by the `eventsByTag` [Persistence Query](http://doc.akka.io/docs/akka/2.4/java/persistence-query.html) implemented by [akka-persistence-cassandra](https://github.com/akka/akka-persistence-cassandra). The tag corresponds to the `tag` defined by the `AggregateEventTag`.
