+++ docs/manual/guide/services/ServiceDescriptors.md
@@ -64,7 +64,7 @@ So far, all of the service call examples we've seen have used strict messages, f
 
 ### Streamed messages
 
-A streamed message is a message of type [`Source`](http://doc.akka.io/japi/akka/2.4.4/akka/stream/javadsl/Source.html).  `Source` is an [Akka streams](http://doc.akka.io/docs/akka/2.4.4/java.html) API that allows asynchronous streaming and handling of messages.  Here's an example streamed service call:
+A streamed message is a message of type [`Source`](http://doc.akka.io/japi/akka/2.4.4/akka/stream/javadsl/Source.html).  `Source` is an [Akka streams](http://doc.akka.io/docs/akka/2.4/java.html) API that allows asynchronous streaming and handling of messages.  Here's an example streamed service call:
 
 @[call-stream](code/docs/services/FirstDescriptor.java)
