+++ persistence-cassandra/javadsl/src/main/scala/com/lightbend/lagom/javadsl/persistence/cassandra/CassandraSession.scala
@@ -1,0 +1,262 @@
+/*
+ * Copyright (C) 2016 Lightbend Inc. <http://www.lightbend.com>
+ */
+package com.lightbend.lagom.javadsl.persistence.cassandra
+
+import javax.inject.Singleton
+import java.util.{ List => JList }
+import java.util.Optional
+import java.util.concurrent.CompletionStage
+
+import scala.annotation.varargs
+
+import akka.stream.javadsl
+import com.datastax.driver.core.BatchStatement
+import com.datastax.driver.core.PreparedStatement
+import com.datastax.driver.core.Row
+import com.datastax.driver.core.Session
+import com.datastax.driver.core.Statement
+import akka.Done
+import akka.NotUsed
+import akka.actor.ActorSystem
+import scala.concurrent.ExecutionContext
+import javax.inject.Inject
+import akka.persistence.cassandra.session.CassandraSessionSettings
+import akka.actor.ExtendedActorSystem
+import akka.event.Logging
+import akka.persistence.cassandra.SessionProvider
+import scala.concurrent.Future
+import akka.persistence.cassandra.CassandraPluginConfig
+
+/**
+ * Data Access Object for Cassandra. The statements are expressed in
+ * <a href="http://docs.datastax.com/en/cql/3.3/cql/cqlIntro.html">Cassandra Query Language</a>
+ * (CQL) syntax.
+ *
+ * The configured keyspace is automatically created if it doesn't already exists. The keyspace
+ * is also set as the current keyspace, i.e. it doesn't have to be qualified in the statements.
+ *
+ * All methods are non-blocking.
+ */
+@Singleton
+final class CassandraSession(system: ActorSystem, settings: CassandraSessionSettings, executionContext: ExecutionContext) {
+
+  @Inject
+  def this(system: ActorSystem) =
+    this(
+      system,
+      settings = CassandraSessionSettings(system.settings.config.getConfig(
+        "lagom.persistence.read-side.cassandra"
+      )),
+      executionContext = system.dispatchers.lookup(system.settings.config.getString(
+        "lagom.persistence.read-side.use-dispatcher"
+      ))
+    )
+
+  private val delegate: akka.persistence.cassandra.session.javadsl.CassandraSession = {
+
+    // FIXME all this can be extracted to core internal and used from both javadsl and scaladsl
+
+    import akka.util.Helpers.Requiring
+    import scala.collection.JavaConverters._
+    import akka.persistence.cassandra.ListenableFutureConverter // implicit asScala conversion
+    val cfg = settings.config
+    val replicationStrategy: String = CassandraPluginConfig.getReplicationStrategy(
+      cfg.getString("replication-strategy"),
+      cfg.getInt("replication-factor"),
+      cfg.getStringList("data-center-replication-factors").asScala
+    )
+
+    val keyspaceAutoCreate: Boolean = cfg.getBoolean("keyspace-autocreate")
+    val keyspace: String = cfg.getString("keyspace").requiring(
+      !keyspaceAutoCreate || _ > "",
+      "'keyspace' configuration must be defined, or use keyspace-autocreate=off"
+    )
+
+    def init(session: Session): Future[Done] = {
+      implicit val ec = executionContext
+      if (keyspaceAutoCreate) {
+        val result1 =
+          session.executeAsync(s"""
+            CREATE KEYSPACE IF NOT EXISTS $keyspace
+            WITH REPLICATION = { 'class' : $replicationStrategy }
+            """).asScala
+        result1.flatMap { _ =>
+          session.executeAsync(s"USE $keyspace;").asScala
+        }.map(_ => Done)
+      } else if (keyspace != "")
+        session.executeAsync(s"USE $keyspace;").asScala.map(_ => Done)
+      else
+        Future.successful(Done)
+    }
+
+    val metricsCategory = "lagom-" + system.name
+    // using the scaladsl API because the init function
+    val scaladslCassandraSession = new akka.persistence.cassandra.session.scaladsl.CassandraSession(
+      system,
+      SessionProvider(system.asInstanceOf[ExtendedActorSystem], settings.config),
+      settings,
+      executionContext,
+      Logging.getLogger(system, this.getClass),
+      metricsCategory,
+      init
+    )
+    new akka.persistence.cassandra.session.javadsl.CassandraSession(scaladslCassandraSession)
+  }
+
+  /**
+   * The `Session` of the underlying
+   * <a href="http://datastax.github.io/java-driver/">Datastax Java Driver</a>.
+   * Can be used in case you need to do something that is not provided by the
+   * API exposed by this class. Be careful to not use blocking calls.
+   */
+  def underlying(): CompletionStage[Session] =
+    delegate.underlying()
+
+  /**
+   * See <a href="http://docs.datastax.com/en/cql/3.3/cql/cql_using/useCreateTableTOC.html">Creating a table</a>.
+   *
+   * The returned `CompletionStage` is completed when the table has been created,
+   * or if the statement fails.
+   */
+  def executeCreateTable(stmt: String): CompletionStage[Done] =
+    delegate.executeCreateTable(stmt)
+
+  /**
+   * Create a `PreparedStatement` that can be bound and used in
+   * `executeWrite` or `select` multiple times.
+   */
+  def prepare(stmt: String): CompletionStage[PreparedStatement] =
+    delegate.prepare(stmt)
+
+  /**
+   * Execute several statements in a batch. First you must [[#prepare]] the
+   * statements and bind its parameters.
+   *
+   * See <a href="http://docs.datastax.com/en/cql/3.3/cql/cql_using/useBatchTOC.html">Batching data insertion and updates</a>.
+   *
+   * The configured write consistency level is used if a specific consistency
+   * level has not been set on the `BatchStatement`.
+   *
+   * The returned `CompletionStage` is completed when the batch has been
+   * successfully executed, or if it fails.
+   */
+  def executeWriteBatch(batch: BatchStatement): CompletionStage[Done] =
+    delegate.executeWriteBatch(batch)
+
+  /**
+   * Execute one statement. First you must [[#prepare]] the
+   * statement and bind its parameters.
+   *
+   * See <a href="http://docs.datastax.com/en/cql/3.3/cql/cql_using/useInsertDataTOC.html">Inserting and updating data</a>.
+   *
+   * The configured write consistency level is used if a specific consistency
+   * level has not been set on the `Statement`.
+   *
+   * The returned `CompletionStage` is completed when the statement has been
+   * successfully executed, or if it fails.
+   */
+  def executeWrite(stmt: Statement): CompletionStage[Done] =
+    delegate.executeWrite(stmt)
+
+  /**
+   * Prepare, bind and execute one statement in one go.
+   *
+   * See <a href="http://docs.datastax.com/en/cql/3.3/cql/cql_using/useInsertDataTOC.html">Inserting and updating data</a>.
+   *
+   * The configured write consistency level is used.
+   *
+   * The returned `CompletionStage` is completed when the statement has been
+   * successfully executed, or if it fails.
+   */
+  @varargs
+  def executeWrite(stmt: String, bindValues: AnyRef*): CompletionStage[Done] =
+    delegate.executeWrite(stmt, bindValues: _*)
+
+  /**
+   * Execute a select statement. First you must [[#prepare]] the
+   * statement and bind its parameters.
+   *
+   * See <a href="http://docs.datastax.com/en/cql/3.3/cql/cql_using/useQueryDataTOC.html">Querying tables</a>.
+   *
+   * The configured read consistency level is used if a specific consistency
+   * level has not been set on the `Statement`.
+   *
+   * You can return this `Source` as a response in a `ServiceCall`
+   * and the elements will be streamed to the client.
+   * Otherwise you have to connect a `Sink` that consumes the messages from
+   * this `Source` and then `run` the stream.
+   */
+  def select(stmt: Statement): javadsl.Source[Row, NotUsed] =
+    delegate.select(stmt)
+
+  /**
+   * Prepare, bind and execute a select statement in one go.
+   *
+   * See <a href="http://docs.datastax.com/en/cql/3.3/cql/cql_using/useQueryDataTOC.html">Querying tables</a>.
+   *
+   * The configured read consistency level is used.
+   *
+   * You can return this `Source` as a response in a `ServiceCall`
+   * and the elements will be streamed to the client.
+   * Otherwise you have to connect a `Sink` that consumes the messages from
+   * this `Source` and then `run` the stream.
+   */
+  @varargs
+  def select(stmt: String, bindValues: AnyRef*): javadsl.Source[Row, NotUsed] =
+    select(stmt, bindValues: _*)
+
+  /**
+   * Execute a select statement. First you must [[#prepare]] the statement and
+   * bind its parameters. Only use this method when you know that the result
+   * is small, e.g. includes a `LIMIT` clause. Otherwise you should use the
+   * `select` method that returns a `Source`.
+   *
+   * The configured read consistency level is used if a specific consistency
+   * level has not been set on the `Statement`.
+   *
+   * The returned `CompletionStage` is completed with the found rows.
+   */
+  def selectAll(stmt: Statement): CompletionStage[JList[Row]] =
+    delegate.selectAll(stmt)
+
+  /**
+   * Prepare, bind and execute a select statement in one go. Only use this method
+   * when you know that the result is small, e.g. includes a `LIMIT` clause.
+   * Otherwise you should use the `select` method that returns a `Source`.
+   *
+   * The configured read consistency level is used.
+   *
+   * The returned `CompletionStage` is completed with the found rows.
+   */
+  @varargs
+  def selectAll(stmt: String, bindValues: AnyRef*): CompletionStage[JList[Row]] =
+    delegate.selectAll(stmt, bindValues: _*)
+
+  /**
+   * Execute a select statement that returns one row. First you must [[#prepare]] the
+   * statement and bind its parameters.
+   *
+   * The configured read consistency level is used if a specific consistency
+   * level has not been set on the `Statement`.
+   *
+   * The returned `CompletionStage` is completed with the first row,
+   * if any.
+   */
+  def selectOne(stmt: Statement): CompletionStage[Optional[Row]] =
+    delegate.selectOne(stmt)
+
+  /**
+   * Prepare, bind and execute a select statement that returns one row.
+   *
+   * The configured read consistency level is used.
+   *
+   * The returned `CompletionStage` is completed with the first row,
+   * if any.
+   */
+  @varargs
+  def selectOne(stmt: String, bindValues: AnyRef*): CompletionStage[Optional[Row]] =
+    delegate.selectOne(stmt, bindValues: _*)
+
+}
+
